build: ds-356 set default value for QUIPUCORDS_APP_PORT
This enables a user to start a working container with just:

    podman run -p 9443:9443 quipucords-ui:latest

Previously the QUIPUCORDS_APP_PORT environment variable was required, and nginx would fail to start and present a very non-obvious error indicating its absence. The variable is still available if you need it but now has a reasonable default value.