refactor(testing): DISCOVERY-523 Add attributes to aid UI tests
The goal is to add ouia ids to ease Camayoc UI testing. Where possible,
ouia id used is the same as id for the element in old UI.

ouiaId attribute is preferred. However, newer Patternfly does not
support it in some elements, or made questionable decisions in other
cases. data-ouia-component-id was used then.

SimpleDropdown interface had to be changed to allow caller to specify
ouia ids of dropdown elements. ouia ids are optional.