ci: drop commit lint action
The commit lint action is more strict than conventional commits. This is causing ALL weekly PRs opened by Konflux/Mintmaker to fail.

This should not be a big issue because the check is optional, but any Github CI failure also prevents Camayoc from running.

This issue is not worth of any developer time.