chore(deps): replace react-table-batteries with its vendored version
For some reason typescript can resolve all types used on the vendored
dependencies, while eslint parser had ~25 false import errors. Maybe
this is related to the fact that the eslint and TS versions in use are
not compatible (there's even a warning about that when running `npm run test:lint`.

Furthermore all vendored code is being ignored from testing coverage.
If not ignored, the vendored could would have plummeted our current
coverage percentage.

Finally, vendoring these libraries resolved the dependency hell with
libraries like @tanstack/react-query (see [1] and [2]). This removes
the need of using the "--legacy-peer-deps" option on npm install during
the production build. This occurred because all conflicts with incompatible
libraries were solved. Because of the resolution of this problem, the option
"overriding" axios package on "package.json" is not required anymore.

Relates to JIRA: DISCOVERY-798
Relates to JIRA: DISCOVERY-1035

[1]: https://github.com/quipucords/quipucords-ui/blob/6f50e1fabf7a1cdf9215884193fe44e1119fa38a/package-lock.json#L18
[2]: https://github.com/quipucords/quipucords-ui/blob/6f50e1fabf7a1cdf9215884193fe44e1119fa38a/package-lock.json#L3318