refactor: remove update-konflux-pipeline target from Makefile
Recently it became clear that pipeline-patcher (which was used on update-konflux-pipeline target) was not enough to properly upgrade
our tekton pipelines. It became clear that using this tool alone represents a risk, so let's remove it.

In the future we might reproduce what MintMaker does by running it locally [1] with MintMaker config [2], which relies on a tool called pipeline-migration-tool to properly handle migrations. Or maybe just on MintMaker to propose updates to our pipelines.

[1]: https://docs.renovatebot.com/modules/platform/local/
[2]: https://github.com/konflux-ci/mintmaker/blob/a3816e76f5e06ba3114731fc34fa14f76d442b6c/config/renovate/renovate.json#L70-L116