test(scans): Properly isolate useScanApi hooks tests
`mockClear()` clears the information about the mock, while keeping the
mock in place. Also, jest mock stacks - you can `spyOn` the same object
twice, and generally the latest spy will be executed, but there are
cases where spy may "fall through" - and some earlier spy will be
executed instead / additionally.

As we were mocking functions in external module and we called `mockClear()`,
we just kept on adding new mocks on top of existing ones. This made
things completely unpredictable, as mock created in one `describe()`
could be executed in completely different `describe()` block.

`mockRestore()` not only clears the information about the mock, but also
reverts the spied function back to the state before spy was created.

Also, `spyOn()` in jest allows to keep track of function invocations and
arguments, but by default *does not mock return value*. The spied
function is executed. That means that we need to *always* chain
`mockResolvedValue`, `mockRejectedValue` or `mockImplementation` on
functions that have side-effects, modify a global state or send HTTP
requests.