Change SimpleDropdown interface to allow parent to decide if it should be opened or not
This is actually the last change made in the journey. As we did all the
other changes and exhausted our other options, the last attempt was to
invert the logic and let parent decide if the dropdown should be opened
or not.

This change is likely not needed in correct final implementation, but it
did help to establish there is also PatternFly bug encountered here.