Ignore this commit
This commit should not exist. Since we changed SimpleDropdown interface
in previous commit, now TypeScript complains about all the places where
SimpleDropdown is used without mandatory parameters. Since the goal was
to push to see how DISCOVERY-819 can be fixed, the easiest way forward
was to just remove other SimpleDropdown references, to silent errors.
Obviously this is not a solution we can actually use.