Change routes to keep reference to element class instead of rendered element
Gemini believes that keeping a reference to rendered element inside routes
causes it to re-render when children components change. Supposedly,
changing the place where element is rendered fixes the issue.

I don't know and that change alone did not bring us much closer to a
solution, but that was one of steps we have taken when we exhausted our
options inside view and Gemini suggested we start modifying parent
components.