Move auth to separate module
At this point Gemini believed that Login page is at fault, but keeping
all components in a tree did not help. Gemini was confident the problem
must be in useGetSetAuth. The first solution (not present here) was to
add a bit of state management. Gemini thinks that setIsAuthorized is
called for each request, and that causes a re-render. Later Gemini
decided there must be a race condition between rendering and requests,
and suggested to move things away from useGetSetAuthApi into a brand new
service. This commit represents a state after implementing that change.

Some intermediate steps could be a better idea than this final version,
as perhaps authentication code runs too often and is not memoized
properly somewhere. But there's a chance this entire Login was a wrong
turn.