Try the hardest to fix a view
This represents a final version of the view code after all the attempts.
It kind of works, in a sense: clicking a button shows a dropdown, and
only clicking anywhere else hides it permanently. However:

- dropdown appears in top left corner of page, instead of below the
  button
- each page rerender hides and displays a dropdown quickly, but slow
  enough to notice a flicker

The changes done here are accumulated over multiple iterations and chat
messages. Initially changes were done here, then Gemini suggested to
move up the component tree (see previous commits), and finally we moved
back here, at what point Gemini decided the problem is in PatternFly
itself. (To be clear: it's highly likely that the problem is in
PatternFly, but there are some changes needed on Quipucords side
anyway).

Initially, Gemini said that the problem is that there's
renderAddCredsButton (invoked as `{renderAddCredsButton()}`) instead of
AddCredsButton (referenced as `<AddCredsButton />`).

Later, Gemini decided that the problem is that components are defined
inside CredentialsListView definition, and therefore the new definition
is created during each rerender. The fix is to move both `AddCredsButton`
and `renderToolbar` outside of `CredentialsListView`. While not a
terrible idea anyway, unfortunately that creates ungodly amount of props
for both components, as they rely heavily on states and hooks defined at
the top of CredentialsListView.

(Also, Gemini was confident that this can only be fixed by moving
components definition outside of `CredentialsListView` and no amount of
`useCallback` or `useMemo` could help.)

Yet later, Gemini suggested to wrap `onRefresh` and `hasSelectedCredentials`
in `useCallback`. Again, probably not a terrible idea.

It also suggested to stop using `renderToolbar` function / component,
but put a body directly in `PageSection`. Gemini generally had a bit of
back and forth on that point, at various points suggesting to keep stuff
inlined (as in final version), or keep it in outside component. It also
suggested to memoize the component or not.

Once that didn't help, Gemini suggested changes contained in other
commits, in this order:

- routes (use class reference instead of element reference)
- <Login> CSS change
- adding state management to Login component
- refactoring `useGetSetAuthApi` a bit
- memoizing various components and functions inside view
- refactoring `useGetSetAuthApi` completely, including an introduction
  of brand new service
- lifting <SimpleDropdown> open state management to parent components
- adding `setTimeout` to `handleDropdownToggle`, to work around a likely bug in
  PatternFly