test(sources): isolate form validation tests
`useSourceForm` contains effect that is called for each change in
`typeValue` (which represents the source type). That effects makes API
request. TypeValue is set (= changed) when `useSourceForm` is first
called. Consequently, form validation tests made stray network requests,
which showed as errors in test output (note that tests were still
passing - but there was a message claiming there was an error).

Mock out axios get requests to properly isolate the tests and hide the
error message.