fix(overview): don't leak icon color change to other views
This fixes a following bug:
1. Have some credentials and sources. Ideally with successful scans.
2. Open Overview page (default after logging in). Refresh page.
3. Open Sources page using left menu

Result: Icon in "Last Connected" column has changed color. It is blue
instead of green.
Refreshing a page on Sources view returns color back to normal.

The root cause is that CSS is split and loaded conditionally as
components are mounted, but it is not unloaded later - so all component
CSS is global, but not necessarily loaded right after refresh.