fix(TypeaheadCheckboxes): don't show "check" twice
PatternFly Menu and Select share some code. Menu items can be selected -
a "check" icon will be displayed on the right end of the item. Select
items may contain <input type=checkbox> to express the same idea.

In PatternFly 6, there seems to be a regression - Select items with
hasCheckbox={true} *also* display a "check" icon on the right, but only
on hover.

Override the default styles so icon on right is never displayed if
there's already an <input> element.